requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
airportsdata==20241001
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
astor==0.8.1
async-timeout==4.0.3
attrs==24.2.0
blake3==1.0.4
cachetools==5.5.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorful==0.5.6
compressed-tensors==0.9.1
datasets==2.21.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.10.0
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.8
importlib_metadata==8.4.0
iniconfig==2.0.0
interegular==0.3.3
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lark==1.2.2
linkify-it-py==2.0.3
llvmlite==0.43.0
lm-format-enforcer==0.10.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mistral_common==1.5.2
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.61.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.2
partial-json-parser==0.2.1.1.post4
pillow==10.4.0
platformdirs==4.3.2
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
proto-plus==1.26.0
protobuf==5.28.0
psutil==6.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.35.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
textual==1.0.0
tiktoken==0.7.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.48.2
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.25.1
vllm==0.7.2
vllm-flash-attn==2.6.1
watchfiles==0.24.0
websockets==13.0.1
wrapt==1.17.2
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.10.0
zipp==3.20.1
